Multi plotter and theming (#20)
This PR allows for proper customization of graphs, providing a mechanism to pass any possible keyword to each of the makie functions. Building on this it also allows to plot multiple sets of data together, although it still needs some adjustment in order to include the credible intervals within the graph.

* Remove hard-coded theming, start multi plot implementation

* Added missing file

* Multiplot sets ranges accounting for all sets given

* Added theming for multiplot

* Finished for now, need to display CI in the future

* Update docs Manifest

* Try removing manifest from docs